+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 ## Version History
 
+#### 2.0.0
+
+* Add autodetection of `pyproject.toml` and `runtime.txt` Python version
+    specifiers.
+
+* Fixes for OSX `grep` incompatibility.
+
+* The `--version` flag is now called `--python`, to make it possible
+    to query the plugin version.
+
 #### 1.0.0
 
-* Initial public release
+* Initial public release.